requirements.txt
Flask==3.1.2
gunicorn==25.0.1
requests==2.32.5
PyYAML==6.0.3
dnspython==2.8.0